+++ app/src/test/java/ch/sbb/polarion/extension/generic/context/CurrentContextConfig.java
@@ -1,0 +1,12 @@
+package ch.sbb.polarion.extension.generic.context;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CurrentContextConfig {
+    String value() default "text-extension";
+}