+++ app/src/test/java/ch/sbb/polarion/extension/generic/fields/CustomFieldEnumConverterTest.java
@@ -81,14 +81,6 @@ public class CustomFieldEnumConverterTest {
     }
 
     @Test
-    void testGetEnumOptionByValue() {
-        FieldMetadata fieldMetadata = ConverterTestUtils.getWorkItemCustomField();
-
-        assertEquals(new EnumOption(YES_NO_ENUM_ID, YES.left()), ConverterTestUtils.process(YES.left(), fieldMetadata));
-        assertEquals(new EnumOption(YES_NO_ENUM_ID, NO.left()), ConverterTestUtils.process(NO.left(), fieldMetadata));
-    }
-
-    @Test
     void testGetEnumOptionByValueIgnoringSpaces() {
         FieldMetadata fieldMetadata = ConverterTestUtils.getWorkItemCustomField();
         // Experimentally proven that Polarion allows to create an enum with spaces at the end of its name
@@ -144,11 +136,6 @@ public class CustomFieldEnumConverterTest {
         assertEquals(new EnumOption(YES_NO_ENUM_ID, NO.left()), ConverterTestUtils.process("Nein", context, fieldMetadata));
 
         assertEquals(new EnumOption(YES_NO_ENUM_ID, YES.left()), ConverterTestUtils.process("yes", context, fieldMetadata));
-
-        //default key/value option identification hidden by the implicit mapping
-        assertThrows(IllegalArgumentException.class,
-                () -> ConverterTestUtils.process("Ja", context, fieldMetadata),
-                "Unsupported value 'Ja' for enum ''");
 
         //just an unknown option
         assertThrows(IllegalArgumentException.class,